fixed #2059: Fix complication errors on Android.

+++ cocos2dx/support/ccUTF8.cpp
@@ -275,7 +275,7 @@ cc_utf8_get_char (const char * p)
 }
 
 
-unsigned short* cc_utf8_to_utf16(const char* str_old, size_t length/* = -1 */, size_t* rUtf16Size/* = NULL */)
+unsigned short* cc_utf8_to_utf16(const char* str_old, int length/* = -1 */, int* rUtf16Size/* = NULL */)
 {
     int len = cc_utf8_strlen(str_old, length);
     if (rUtf16Size != NULL) {
